release workflow
workflow that will build and publish (to pypi) python packages, build documentation
and "deploy" it in branch gh-pages.
if no version is specified to the workflow, it will autobump semver based on existing tags.
creates a release tag.

documentation will not be deployed
on readthedocs, removing all dependencies
for that.

docs/licenses.md is now only a template in the main branch, will be generated when
building the documentation.