added new grizzly logo
logo in README.md will not work until repository is set to public. needs
to be included in this way so it will work both on gh-page hosted documentation AND
when looking at it on github.com

updated release workflow to manually
copy docs/assets to the build/html
directory for them to be included in the
published site.

cudos to @boffman for the logo!