improved base for adding different types of tasks
before it was possible to add RequestContext or float, where the later
resulted in a sleep.

improved the base so that tasks added to a scenario is of type GrizzlyTask.

each task inherits this type, and adds it specific implementation.

this made it quite simple to add a third type
of task: PrintTask, which can be used to
print a message (more specifically containing variables) and the message
will be rendered with variables values it have
at the time of execution. useful for
visualizing variable values.