new task to parse data
makes most sense if one has saved an object or part
of an document in a save handler, and wants
to extract more specific information from that result.
E.g., saving a document with AtomicMessageQueue, and then
extracts 2 or more values to variables from that document.

Also... got feeling and fixed two TODO's, to allow user class
implementation and task class implementations outside of
grizzly namespace.