update dependecies, stage 3
use built in cache in setup-python in github workflows.
does not work with act, so adding it as an conditional step.

added more python versions for tests.
pip-generate-requirements.bash runs pip-compile in a docker container with the lowest python version
for the project, as specified in setup.py

added backport.zoneinfo for python versions < 3.9 in setup.py