use locust.contrib.fasthttp.FastHttpSession instead of default client
gave up trying to get async requests to work with the already implemented
`request` method in `GrizzlyUser`. `requests` (RestApiUser) has a problem with
using the same session in multiple threads (see https://github.com/psf/requests/issues/1871).

this was causing the requests to be more or less sequential.

this is solved by implementing an "interface", AsyncRequests, which requires
the user to implement method `async_request`, which then `AsyncRequestGroup`
calls. this implies that only users that implements this interface can be used.

in RestApiUser, `async_request` uses `locust.contrib.fasthttp.FastHttpSession`
to achive this, and creates one session per call to `async_request`, so that
each call is isolated and not dependent on other calls.

no test coverage, yet.