fixed flow to work with bug fix AAD had implemented
state sCtx and some other stuff is now updated between requests in the auth flow, so we need to update the state as well.

also, implementation of auth via cookies, by providing `initialize_uri` which should be called to start the auth flow, which in the ended
returns a cookie on the callback uri. this cookie is saved for `GrizzlyHttpAuthClient` implementations, and is used for all requests
done by them.