handling of SIGINT/SIGTERM to gracefully stop test
this feature required tasks to run in a new greenlet, so it was possible to
abort them by throwing an exception in it to stop any "long going" executions (such as sleep, waiting for messages from async-messaged etc).

when a test is aborted, it will stop IteratorScenario from continuing until all tasks in an scenario iteration is complete.

handling it gracefully will then also mean that all on_stop logic will run
and any resources that should be cleaned up is indeed cleaned up.

scenarios will get status "skipped" when aborted, but the feature will still have status failed.

instead of "Ended" timestamp, it will be "Aborted". And if a test is not aborted it will be "Finished".