implementation of #229
new decorator for filter implementations. the decorated function will be added to the jinja2 filters, with the same name as the function.

removed direct usage of `jinja2.Template`, since it creates a new Environment instance every time, and has a cache of max 10.

added documentation, example implementation and updated tests.