fixed problems related to StopUser and RestartScenario
in the case of StopUser and spawning isn't complete, locust will start a new user even though it shouldn't.
easiest way to get around this is to wait for the scenario to stop until spawning is complete (by listening on the event) to trick locust dispatch logic into that no new user should be spawned.

in the case of RestartScenario, only the task index was reset, but if there where following tasks in the task queue, they would still be executed.
not sure, but to be sure the task queue is now cleared before scenario is
restarted.