refactoring regarding `GrizzlyContextScenario` references for users
each user should have a copy of the `GrizzlyContextScenario` that they are part of. this is to make it possible to, per user, change some aspects of the scenario during runtime without affecting other users.

references to the scenario from `GrizzlyTask` has been removed, in what ever context a task is executed, that context knows about the scenario.

this caused a lot of small changes, mainly in unit tests. at the same time, some fixtures was rewritten to make them simpler and more straight forward to use.