refactoring regarding `GrizzlyContextScenario` references for users (#241)
* fixed problems related to StopUser and RestartScenario

in the case of StopUser and spawning isn't complete, locust will start a new user even though it shouldn't.
easiest way to get around this is to wait for the scenario to stop until spawning is complete (by listening on the event) to trick locust dispatch logic into that no new user should be spawned.

in the case of RestartScenario, only the task index was reset, but if there where following tasks in the task queue, they would still be executed.
not sure, but to be sure the task queue is now cleared before scenario is
restarted.

* scenario instances could be initiated *after* spawning_complete event has been fired

this caused the scenario to wait for ever. use `grizzly.state.spawning_complete` instead.

* additional error logging when task fails

* log changes of failure exception

* refactoring regarding `GrizzlyContextScenario` references for users

each user should have a copy of the `GrizzlyContextScenario` that they are part of. this is to make it possible to, per user, change some aspects of the scenario during runtime without affecting other users.

references to the scenario from `GrizzlyTask` has been removed, in what ever context a task is executed, that context knows about the scenario.

this caused a lot of small changes, mainly in unit tests. at the same time, some fixtures was rewritten to make them simpler and more straight forward to use.

* do not use `dataclasses.replace` to copy an instance of a dataclass object

instead use `copy.copy`.

* response handlers should only dump matches as json if explicitly specified