refactoring of handling RequestTask (#246)
* refactoring of handling RequestTask

rendering a RequestTask now returns a copy of the "template" RequestTask, where all attributes has been rendered (closes #153).

all logic around firing request events, response events etc. har handled in the final
decorated method `GrizzlyUser.request`. all users has an `request_impl` which is
specific to the protocol/transport method that user implements. this make user
implementations cleaner.

a user does not need to know about the scenario it is executing, hence removed
`parent` argument. this caused that `grizzly.auth` needed to be rewritten again,
but for the better. a `ClientTask` now know about the locust environment and
the context for the scenario it is executing.

* due to changes of `parent` not being part of the request

the HttpTask did not get variables.
update the client task `_context` with `parent.user._context` every time the task is executing.
might be another iteration that has received new variables.

---------

Co-authored-by: boffman <686610+boffman@users.noreply.github.com>