async-messaged logging refatored (#287)
removed `ThreadLogger` class while implementing the fix for #283.
that class was created as a possible work-around where tests running
distributed would fail at after a while. turned out it was due to a
buffer becoming full (since it wasn't read), but this logger wrapper
was never removed.

metadata should not be it's own instance variable, it should read
`self._context['metadata']`, so removed the instance variable and
created property getter/setter that wraps around `_context` instead.