cache existing authentication tokens when chaning username+password
when changing `auth.user.username` and `auth.user.password` during runtime, any existing authorization tokens should be cached, so they can be re-used if changed back to the original user.

this is to avoid having to authenticate everytime a user is switched.

tokens are only refresh when the original `refresh_time` has expired.