custom dispatcher logic to be able to isolate user types to workers (#302)
* adaptations for the new dispatcher in locust

* additional logic for handling new dispatcher implementation

* allow undeclared variables, if the variable is part of a "is defined" statement

* `_parse_templates` should return a set of variables that are part of define tests

this is so `initialize_testdata` doesn't have to complain about a variable that are part of a define test not being used in any templates (since it's ignored when finding variables in all templates).

* session_started included in errorr message if response is 401

* info logging message when token is updated

* if master runner fails to start, send quit to workers

* when one token is refreshed due to session has been going on for `refresh_time` the `__cached_auth__` dict has to be cleared, so any other cached auth tokens are refreshed.

* JsonTransformer does not like empty strings

handle them as `"null"` instead, so we don't get an additional exception.

* new user step that sets user type, host, fixed count

and optionally sticky tag.
deprecates/removes grizzly.steps.scenario.shape.

* debug mq on_start problems

* better handling of variable names that is used in multiple scenarios

this usually comes to light when including steps from different feature files.

* wrap step exceptions in context

continue if there are `AssertionError`, used for initialization problems, so we can present a summary in `after_feature`.

changed a bunch of initialization exceptions to `AssertionError`.
changed a couple of exception messages.

* do not persist keystore

* need to be 3.9 compatible with type annotations

* handle slices on jinja2 GetItem ast objects

* handle jinja2 Call ast nodes

e.g. `AtomicRandomString.value.replace`.

* adapt to possibility to "plugin" own dispatcher logic in locust

* fixed problem in logic so unit tests pass

must profile, seems a bit slow.

* adaptations to latest locust PR

* ruff lint fixes

* save optional payload/metadata expression values

if no match, use specified default value.

* use official locust release instead of branch

fixed pyright/mypy clashes. do not use `mypy_extensions`.

* code quality check fixes

* update github actions to avoid node 16 deprecation warnings

* debug failing e2e tests

* handle MQMIError exceptions for put operations

suppress exceptions that happens when trying to close a queue, might not be open.

* new task: WriteFileTask

write content to file.