implementation of dry run
executes basically everything, up until it is time to starts locust workers.

makes it easier to troubleshoot a new feature-file to make sure that behave phase of grizzly at least will work, and that a test case is correctly setup.

this required moving `initialize_testdata` from deep within locust as a library logic.

some basic preparation for #281, as to have a place to store the scenario class type name in the context.