servicebus client task subscription names
subscription names cannot be longer than 50 characters. if a rule is to be created "on-the-fly", a unique identifier which is 8 characters is suffixed to the name, which in case grizzly will only allow at max a name of 42 characters.

the identifier used to be the id of the user instance executing the task, however that might not be sufficient. now it is the first 8 characters of the sha256 hash of user instance id, hostname and timestamp.