credential support for `ServiceBus*` and `BlobStorage*` resources (#318)
* fixed bugs related to "app" authentication flow

* better handling of exceptions that occurs in `on_start`

* support for credential in `BlobStorageUser`

* move AzureAadCredential to grizzly_extras

this is needed so that async-messaged can use credential when creating ServiceBusClient.

* support in async-messaged for ServiceBus credentials

* support for credential in `ServiceBusUser`

* support for credential in `BlobStorageClientTask`

also removed support for specifying `Container` in query string, it must now be specified in the URL path.

updated documentation, and tests.

* support for credential in `ServiceBusClientTask`

* no gevent stuff in grizzly_extras

* align task settings in BlobStorageClientTask with how they are specified in ServiceBusClientTask

* fixed quoting bug for expression in ServiceBusClientTask

regex failed when having expression as the last path element and using credential for authentication.

* servicebus client task subscription names

subscription names cannot be longer than 50 characters. if a rule is to be created "on-the-fly", a unique identifier which is 8 characters is suffixed to the name, which in case grizzly will only allow at max a name of 42 characters.

the identifier used to be the id of the user instance executing the task, however that might not be sufficient. now it is the first 8 characters of the sha256 hash of user instance id, hostname and timestamp.

* CI test fixes

and a unbound variable reference fix.

* fixed broken `initialize_uri` flow

seems like microsoft changed some check related to which headers has been set. without the changes `user auth request 5` would fail with 404.