jsonpath filter one of, `=|` (#325)
* support for `or` or, "one of" in expressions for service bus messages

makes it possible to find a message based on a sequence of values, where the one specific expression on the payload must be in the specified sequence.

* in order to allow `|=` in possible argument values, additional checks is needed since `|` sometimes is used to seperate values from arguments.

* log request that failed to be rendered

* when rendering a template that is the string representation of an json object/list, we must escape any non-json compatible use of characters in side variable values.

* bump to setuptools-scm==8.1.0

git seems to have changed the format of format modifier `%cI`, which uses `Z` as short for `+00:00`. this short is not supported by `datetime.fromisoformat` until python >= 3.11. which causes problems for anything less.