fix for issue #323
start `router` in seperate process, which in turn runs a `ThreadPoolExecutor`, where each worker is a seperate thread.

install the signal handler and share a `threading.Event` with all processes/threads, that if set would abort their respective I/O loops. the signal handler will set the event in case of SIGTERM.

in the main loop, wait for the event to be set, then terminate the process with the `ThreadPoolExecutor`.

the router/frontend will check which threads has stopped by their own, and the others it will send a "abort" message to their clients.

the abort message will result in `AsyncMessageAbort` exception, which is handled by users and client tasks in the form of a `StopUser`.