fix for "Aborting test can fail when running distributed with MQ" (#333)
* reformat behave exception error messages

* rewrite of async-messaged worker

still not quite done with forcing the process to stop when receiving sigint.

* fix for issue #323

start `router` in seperate process, which in turn runs a `ThreadPoolExecutor`, where each worker is a seperate thread.

install the signal handler and share a `threading.Event` with all processes/threads, that if set would abort their respective I/O loops. the signal handler will set the event in case of SIGTERM.

in the main loop, wait for the event to be set, then terminate the process with the `ThreadPoolExecutor`.

the router/frontend will check which threads has stopped by their own, and the others it will send a "abort" message to their clients.

the abort message will result in `AsyncMessageAbort` exception, which is handled by users and client tasks in the form of a `StopUser`.

* py3.12 compatible rmtree wrapper

* fixed unit tests

* bump static python version from 3.11 to 3.12 in workflows

* fix fixed unit tests

* refactoring so that each scenario has their own jinja2 Environment

this is done in preparation of allow same variable names in all scenarios, but them still beeing separated (except for Atomic* variables).

first step though will be the possibility to reuse request templates between scenarios, which will be needed when scenarios includes steps from other scenarios `{% scenario ... %}`.