`IotHubUser` handle C2D messages as `GET`/`RECEIVE`
messages are added to the keystore with key `queue::<device id>` when they are received. the actual grizzly request handling for it is done by waiting for a message in the keystore for that key.

this is so it can run distributed, and it does not matter which client connection actually receives the message.