TaskTimeoutError, make a task timeout if it is running for too long
possibility to add a timeout a task, and grizzly will make sure it does not run longer than that. If it does `grizzly.exceptions.TaskTimeoutError` will be raised, which then can be handled by the failure handling ("When a task fails....").

for `RequestTask` it is possible to set the timeout via pipe-argument `timeout` (`.... | timeout=<seconds>`).

it is also possible to set method and name for a task via `grizzlytask.metadata`, these two will be used in the error summary if a task takes to long to execute.