improved documentation and logic
key for mapping to an actual timer now includes the timer name. so in the case of knowing the actual timer name, then tid + version does not need to be unique.

when starting a timer a mapping between tid + version and the name is added.

so when stopping a timer with an unknown name, tid + version does indeed need to be unique, otherwise an error will be logged.