makes asynchronous timers more asynchronous
due to race conditions for fast timers, and where they are started/stopped and that worker sends its report to the master, there could be timers being stopped before they have been started.

this fixes it, by tracking a timer with a unique id, and it is possible to create and instance with either start or stop timestamp. when both are present, the timer is completed and "request" statistics generated.